updated search
